backend/requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tk==0.1.0
tzdata==2024.2
Werkzeug==3.1.3
flask
flask-sqlalchemy
flask-cors
flask-migrate
reportlab
python-docx==1.1.0
psycopg2-binary>=2.9.11
openpyxl>=3.0.0
xlwt>=1.3.0
PyYAML>=6.0.2
